requirements.txt
alembic==1.4.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mysql-connector==2.1.7
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1
